refactor: 분산된 mock data 하나로 취합
