--- src/app/(pages)/apply/_constants/applyJobMockData.ts
@@ -1,11 +1,0 @@
-export type ApplyJobPosting = {
-  title: string;
-  companyName: string;
-  jobField: string;
-};
-
-export const applyJobMockData: ApplyJobPosting = {
-  title: '[포스코] 2026년 하계 스틸캠프(체험형 인턴십) 인턴 모집',
-  companyName: '포스코',
-  jobField: 'UXUI 디자이너',
-};
